services/search_and_browse_service/requirements.txt
flask
redis
hiredis
pymysql
requests
python-dateutil
awsiotsdk 